feat: Error handling for node process
- Added error handling code for node process
- Separated the index.js and the app.js
